// UI/src/app/model/group.ts
/**
 * TodoApi
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { TodoItem } from './todoItem';

export interface Group { 
    id?: number;
    name?: string;
    isDefault?: boolean;
    modifiedDate?: Date;
    createdDate?: Date;
    items?: Array<TodoItem>;
}